Move `incident_history` related codes to `history.go` file
